+++ madmom/features/onsets.py
@@ -650,7 +650,7 @@ class SpectralOnsetProcessor(Processor):
         return g
 
 
-# class for detecting onsets with RNNs
+# classes for detecting onsets with NNs
 class RNNOnsetProcessor(SequentialProcessor):
     """
     Processor to get a onset activation function from multiple RNNs.
@@ -660,6 +660,24 @@ class RNNOnsetProcessor(SequentialProcessor):
     online : bool, optional
         Choose networks suitable for online onset detection, i.e. use
         unidirectional RNNs.
+
+    Notes
+    -----
+    This class uses either uni- or bi-directional RNNs. Contrary to [1], it
+    uses simple tanh units as in [2]. Also the input representations changed
+    to use logarithmically filtered and scaled spectrograms.
+
+    References
+    ----------
+    .. [1] "Universal Onset Detection with bidirectional Long Short-Term Memory
+            Neural Networks"
+           Florian Eyben, Sebastian Böck, Björn Schuller and Alex Graves.
+           Proceedings of the 11th International Society for Music Information
+           Retrieval Conference (ISMIR), 2010.
+    .. [2] "Online Real-time Onset Detection with Recurrent Neural Networks"
+           Sebastian Böck, Andreas Arzt, Florian Krebs and Markus Schedl.
+           Proceedings of the 15th International Conference on Digital Audio
+           Effects (DAFx), 2012.
 
     """
 
@@ -702,6 +720,67 @@ class RNNOnsetProcessor(SequentialProcessor):
 
         # instantiate a SequentialProcessor
         super(RNNOnsetProcessor, self).__init__((pre_processor, nn))
+
+
+# must be a top-level function to be pickle-able
+def _cnn_onset_processor_pad(data):
+    """Pad the data by repeating the first and last frame 7 times."""
+    pad_start = np.repeat(data[:1], 7, axis=0)
+    pad_stop = np.repeat(data[-1:], 7, axis=0)
+    return np.concatenate((pad_start, data, pad_stop))
+
+
+class CNNOnsetProcessor(SequentialProcessor):
+    """
+    Processor to get a onset activation function from a CNN.
+
+    References
+    ----------
+    .. [1] "Musical Onset Detection with Convolutional Neural Networks"
+           Jan Schlüter and Sebastian Böck.
+           Proceedings of the 6th International Workshop on Machine Learning
+           and Music, 2013.
+
+    Notes
+    -----
+    The implementation follows as closely as possible the original one, but
+    part of the signal pre-processing differs in minor aspects, so results can
+    differ slightly, too.
+
+    """
+
+    def __init__(self, **kwargs):
+        # pylint: disable=unused-argument
+        from ..audio.signal import SignalProcessor, FramedSignalProcessor
+        from ..audio.filters import MelFilterbank
+        from ..audio.spectrogram import (FilteredSpectrogramProcessor,
+                                         LogarithmicSpectrogramProcessor)
+        from ..models import ONSETS_CNN
+        from ..ml.nn import NeuralNetwork
+
+        # define pre-processing chain
+        sig = SignalProcessor(num_channels=1, sample_rate=44100)
+        # process the multi-resolution spec in parallel
+        multi = ParallelProcessor([])
+        for frame_size in [2048, 1024, 4096]:
+            frames = FramedSignalProcessor(frame_size=frame_size, fps=100)
+            filt = FilteredSpectrogramProcessor(
+                filterbank=MelFilterbank, num_bands=80, fmin=27.5, fmax=16000,
+                norm_filters=True, unique_filters=False)
+            spec = LogarithmicSpectrogramProcessor(log=np.log, add=0)
+            # process each frame size with spec and diff sequentially
+            multi.append(SequentialProcessor((frames, filt, spec)))
+        # stack the features (in depth) and pad at beginning and end
+        stack = np.dstack
+        pad = _cnn_onset_processor_pad
+        # pre-processes everything sequentially
+        pre_processor = SequentialProcessor((sig, multi, stack, pad))
+
+        # process the pre-processed signal with a NN ensemble
+        nn = NeuralNetwork.load(ONSETS_CNN[0])
+
+        # instantiate a SequentialProcessor
+        super(CNNOnsetProcessor, self).__init__((pre_processor, nn))
 
 
 # universal peak-picking method