+++ tests/test_ml_nn.py
@@ -63,3 +63,50 @@ class TestNeuralNetworkClass(unittest.TestCase):
         result = rnn.process(data)
         self.assertTrue(np.allclose(result, [0.0815198, 0.24451593,
                                              0.08786312, 0.01776425]))
+
+    def test_cnn(self):
+        cnn = NeuralNetwork.load(ONSETS_CNN[0])
+        data = np.zeros((19, 80, 3))
+        data[10] = 1.
+        result = cnn.process(data)
+        self.assertTrue(np.allclose(result, [0.0021432, 0.02647826, 0.92750794,
+                                             0.84207922, 0.21631248]))
+
+
+class TestBatchNormLayerClass(unittest.TestCase):
+
+    IN = np.array([[[0.32400414, 0.31483042],
+                    [0.38269293, 0.04822304],
+                    [0.03791266, 0.34776369]],
+                   [[0.87113619, 0.62172854],
+                    [0.87353969, 0.92837042],
+                    [0.70359915, 0.49917081]],
+                   [[0.42643583, 0.74653631],
+                    [0.08519834, 0.35423595],
+                    [0.34863797, 0.44895086]]])
+
+    BN_PARAMS = [np.array([-0.00098404, 0.00185387]),
+                 np.array([-0.0068268, 0.0068859]),
+                 np.array([-0.00289366, 0.00742069]),
+                 np.array([-0.00177374, -0.00444383])]
+
+    OUT = np.array([[[-0.00098008, 0.00184446],
+                     [-0.00097937, 0.00185262],
+                     [-0.00098355, 0.00184345]],
+                    [[-0.00097346, 0.00183507],
+                     [-0.00097343, 0.00182569],
+                     [-0.00097549, 0.00183882]],
+                    [[-0.00097884, 0.00183125],
+                     [-0.00098297, 0.00184326],
+                     [-0.00097978, 0.00184036]]])
+
+    def test_batch_norm(self):
+        params = TestBatchNormLayerClass.BN_PARAMS
+        x = TestBatchNormLayerClass.IN
+        y_true = TestBatchNormLayerClass.OUT
+        bnl = layers.BatchNormLayer(
+            params[0], params[1], params[2], params[3], activations.linear)
+
+        y = bnl.activate(x)
+
+        self.assertTrue(np.allclose(y, y_true))