+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 0.14.dev
 New features:
 
 * RNN/DBN downbeat tracking (#130)
+* Covolutional neural networks and CNN onset detection (#133)
 * Linear-Chain Conditional Random Field implementation (#144)
 * DNN based chroma vector extraction (#148)
 * CRF chord recognition using DNN based chroma vectors (#148)